+++ R/model_identify_variables.R
@@ -7,8 +7,10 @@
 #' * `term`: coefficients of the model
 #' * `variable`: the corresponding variable
 #' * `var_class`: class of the variable (cf. [stats::.MFclass()])
-#' * `var_type`: `"continuous"`, `"categorical"`, `"intercept"`
+#' * `var_type`: `"continuous"`, `"dichotomous"` (categorical variable with 2 levels),
+#'   `"categorical"` (categorical variable with 3 or more levels), `"intercept"`
 #'   or `"interaction"`
+#' * `var_nlevels`: number of original levels for categorical variables
 #' @export
 #' @family model_helpers
 #' @seealso [tidy_identify_variables()]
@@ -38,6 +40,20 @@ model_identify_variables <- function(model) {
 #' @export
 model_identify_variables.default <- function(model) {
   model_matrix <- model_get_model_matrix(model)
+
+  if (is.null(model_matrix)) {
+    # return an empty tibble
+    return(
+      dplyr::tibble(
+        variable = NA_character_,
+        var_class = NA_character_,
+        var_type = NA_character_,
+        var_nlevels = NA_integer_
+      ) %>%
+        dplyr::filter(FALSE)
+    )
+  }
+
   assign <- attr(model_matrix, "assign")
   assign[assign == 0] <- NA
   model_terms <- stats::terms(model)
@@ -60,6 +76,11 @@ model_identify_variables.default <- function(model) {
         dplyr::select(.data$variable, .data$var_class),
       by = "variable"
     ) %>%
+    dplyr::left_join(
+      model_get_nlevels(model),
+      by = "variable"
+    ) %>%
+
     .compute_var_type()
 }
 
@@ -74,11 +95,17 @@ model_identify_variables.lavaan <- function(model) {
     dplyr::left_join(
       tibble::tibble(
         variable = model@Data@ov$name,
-        var_class = model@Data@ov$type
+        var_class = model@Data@ov$type,
+        var_nlevels = model@Data@ov$nlev
       ),
       by = "variable"
     ) %>%
     dplyr::mutate(
+      var_nlevels = dplyr::if_else(
+        .data$var_nlevels == 0,
+        NA_integer_,
+        .data$var_nlevels
+      ),
       var_class = dplyr::if_else(
         .data$var_class == "ordered",
         "factor",
@@ -93,11 +120,13 @@ model_identify_variables.lavaan <- function(model) {
 ## model_identify_variables() helpers --------------------------
 
 .compute_var_type <- function(x) {
+  cat_classes <- c("factor", "character", "logical")
   x %>%
     dplyr::mutate(
       var_type = dplyr::case_when(
         is.na(.data$variable) ~ "intercept",
-        .data$var_class %in% c("factor", "character", "logical") ~ "categorical",
+        .data$var_class %in% cat_classes & .data$var_nlevels <= 2 ~ "dichotomous",
+        .data$var_class %in% cat_classes ~ "categorical",
         !is.na(.data$var_class) ~ "continuous",
         is.na(.data$var_class) & stringr::str_detect(.data$variable, ":") ~ "interaction"
       )