+++ R/tidy_identify_variables.R
@@ -1,13 +1,22 @@
 #' Identify the variable corresponding to each model coefficient
 #'
 #' `tidy_identify_variables()` will add to the tidy tibble
-#' three additional columns: `variable`, `var_class` and `var_type`.
+#' three additional columns: `variable`, `var_class`, `var_type` and `var_nlevels`.
 #'
 #' It will also identify interaction terms and intercept(s).
-#' `var_type` could be `"continuous"`, `"categorical"`, `"intercept"`
-#' or `"interaction"`. Will be equal to `"unknown"` in the rare cases
-#' where `tidy_identify_variables()` will fail to identify the list
-#' of variables.
+#'
+#' `var_type` could be:
+#'
+#' * `"continuous"`,
+#' * `"dichotomous"` (categorical variable with 2 levels),
+#' * `"categorical"` (categorical variable with 3 levels or more),
+#' * `"intercept"`
+#' * `"interaction"`
+#' * `"unknown"` in the rare cases where `tidy_identify_variables()`
+#'   will fail to identify the list of variables
+#'
+#' For dichotomous and categorical variables, `var_nlevels` corresponds to the number
+#' of original levels in the corresponding variables.
 #' @param x a tidy tibble
 #' @param model the corresponding model, if not attached to `x`
 #' @inheritParams tidy_plus_plus
@@ -65,8 +74,9 @@ tidy_identify_variables <- function(x, model = tidy_get_model(x),
     if (!quiet)
       usethis::ui_oops(paste0(
         "Unable to identify the list of variables.\n\n",
-        "This is usually due to an error calling {usethis::ui_code('stats::model.frame(x)')}.\n",
-        "It could be the case if that type of model does not implement this method.\n",
+        "This is usually due to an error calling {usethis::ui_code('stats::model.frame(x)')}",
+        "or {usethis::ui_code('stats::model.matrix(x)')}.\n",
+        "It could be the case if that type of model does not implement these methods.\n",
         "Rarely, this error may occur if the model object was created within\na ",
         "functional programming framework (e.g. using {usethis::ui_code('lappy()')}, ",
         "{usethis::ui_code('purrr::map()')}, etc.)."
@@ -76,7 +86,8 @@ tidy_identify_variables <- function(x, model = tidy_get_model(x),
       dplyr::mutate(
         variable = NA_character_,
         var_class = NA_integer_,
-        var_type = "unknown"
+        var_type = "unknown",
+        var_nlevels = NA_integer_
       ) %>%
       tidy_attach_model(model = model, .attributes = .attributes)
   }