+++ tests/testthat/test-identify_variables.R
@@ -2,24 +2,28 @@ library(survival)
 library(gtsummary)
 
 test_that("tidy_identify_variables() works for common models", {
-  mod <- glm(response ~ age + grade * trt, gtsummary::trial, family = binomial)
+  mod <- glm(response ~ age + grade * trt + death, gtsummary::trial, family = binomial)
   res <- mod %>%
     tidy_and_attach() %>%
     tidy_identify_variables()
   expect_equivalent(
     res$variable,
-    c(NA, "age", "grade", "grade", "trt", "grade:trt", "grade:trt")
+    c(NA, "age", "grade", "grade", "trt", "death", "grade:trt", "grade:trt")
   )
   expect_equivalent(
     res$var_class,
-    c(NA, "numeric", "factor", "factor", "character", NA, NA)
+    c(NA, "numeric", "factor", "factor", "character", "integer", NA, NA)
   )
   expect_equivalent(
     res$var_type,
     c(
-      "intercept", "continuous", "categorical", "categorical", "categorical",
-      "interaction", "interaction"
+      "intercept", "continuous", "categorical", "categorical", "dichotomous",
+      "continuous", "interaction", "interaction"
     )
+  )
+  expect_equivalent(
+    res$var_nlevels,
+    c(NA, NA, 3L, 3L, 2L, NA, NA, NA)
   )
 })
 
@@ -135,7 +139,7 @@ test_that("tidy_identify_variables() works with variables having non standard na
   )
   expect_equivalent(
     res$var_type,
-    c("categorical", "categorical", "categorical", "categorical")
+    c("dichotomous", "categorical", "categorical", "categorical")
   )
 })